requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cymem==2.0.6
Django==4.0.1
finnhub-python==2.4.8
idna==3.3
importlib-metadata==4.10.1
Jinja2==3.0.3
langcodes==3.3.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
murmurhash==1.0.6
numpy==1.22.1
packaging==21.3
pandas==1.4.0
pathy==0.6.1
preshed==3.0.6
pycodestyle==2.8.0
pydantic==1.8.2
PyJWT==2.3.0
pyngrok==5.1.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==3.13
regex==2022.1.18
requests==2.27.1
six==1.16.0
smart-open==5.2.1
spacy-loggers==1.0.1
sqlparse==0.4.2
srsly==2.4.2
thinc==8.0.13
toml==0.10.2
tqdm==4.62.3
twilio==7.5.0
typer==0.4.0
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
wasabi==0.9.0
websocket-client==1.2.3
websockets==10.1
zipp==3.7.0
